No desync AI
AI is now running synced, in the sense that end-turn in not send from
within the AI, but instead from the script_engine, when the AI runs out
of scope. Also end-turn now triggers a 10 second delay, such that the AI
does not start running again, before the enemy has played his turn.
(Previously is was possible for the AI to run for something like 1
second (typically until first delay call)).